requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
discord.py==2.3.2
aiohttp==3.9.1
